refactor: #139 WatchViewModel 분리 완료
- iOS PlayerViewModel 패턴 완전 적용
- 클로저 기반 NotificationCenter 사용 (@objc 제거)
- 기능별 Extension 파일로 분리
- weak self 캡처로 메모리 안전성 확보
